Add a interface for the AudioEnvironment.
